pyproject.toml
[tool.poetry]
name = "django-example"
version = "0.1.0"
description = ""
authors = ["greedWizard <skavik46111@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
psycopg2 = "^2.9.9"
django-environ = "^0.11.2"
django-ninja = "^1.1.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
pre-commit = "^3.6.0"
faker = "^22.6.0"
pytest = "^8.0.0"
pytest-django = "^4.8.0"
factory-boy = "^3.3.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.project.settings.local"


[tool.isort]
multi_line_output=3
include_trailing_comma=true
line_length=120
lines_after_imports=2
lines_between_sections=1
lines_between_types=0
force_alphabetical_sort_within_sections=true
ensure_newline_before_comments=true
combine_as_imports=true
combine_star=true
skip_gitignore=true
balanced_wrapping=true
force_grid_wrap=2
src_paths=["core", ]
skip_glob=[
    ".git",
    ".idea",
    ".vscode",
    "venv",
    ".venv",
]
known_django=["django","ninja"]
known_first_party=["core"]
sections=[
    "FUTURE",
    "STDLIB",
    "DJANGO",
    "THIRDPARTY",
    "FIRSTPARTY",
    "LOCALFOLDER",
]